requirements.txt
discord.py[voice]
asyncio
aiohttp
pytz
